backend/requirements.txt
fastapi
uvicorn
qdrant-client
torch
torchvision
transformers
pillow
google-genai
python-multipart
SQLAlchemy
psycopg2-binary
